+++ src/ast/tokenization/append_node.c
@@ -1,61 +1,61 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   append_word_node.c                                 :+:      :+:    :+:   */
+/*   append_node.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: jhor <jhor@student.42kl.edu.my>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/10/22 16:49:26 by jhor              #+#    #+#             */
-/*   Updated: 2025/10/23 20:41:42 by jhor             ###   ########.fr       */
+/*   Created: 2025/10/22 16:48:15 by jhor              #+#    #+#             */
+/*   Updated: 2025/11/12 15:41:29 by jhor             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../minishell.h"
+#include "../../../minishell.h"
 
-void	free_append_word(t_token *head, t_token *temp)
+t_token	*init_node(t_token *token)
 {
-	free_tokens(head);
-	free_tokens(temp);
+	token->token = 0;
+	token->lexeme = NULL;
+	token->next = NULL;
+	return (token);
 }
 
-char	*word_malloc(t_token *token, char *start, size_t *n_start, size_t *end)
+char	*lexeme_malloc(t_token *token, size_t *i, char *start, size_t *end)
 {
-	size_t	i;
-
-	i = 0;
-	while (*n_start < *end && start[*n_start])
+	while (*i < (*end) && start[*i])
 	{
-		token->lexeme[i++] = start[*n_start];
-		(*n_start)++;
+		token->lexeme[*i] = start[*i];
+		(*i)++;
 	}
-	token->lexeme[i] = '\0';
+	token->lexeme[*i] = '\0';
 	return (token->lexeme);
 }
 
-t_token	*append_word_node(t_token *hd, char *srt, size_t n_srt, size_t end)
+t_token	*append_node(t_token *head, char *start, size_t end)
 {
 	t_token	*temp;
+	size_t	i;
 	t_token	*trave;
 
-	temp = NULL;
+	i = 0;
 	temp = malloc(sizeof(t_token));
 	if (!temp)
 	{
-		free_append_word(hd, temp);
+		free_append_word(head, temp);
 		return (NULL);
 	}
 	temp = init_node(temp);
-	temp->lexeme = malloc(sizeof(char) * (end + 1));
+	temp->lexeme = malloc(sizeof(char) * ((end) + 1));
 	if (!temp->lexeme)
 	{
-		free_append_word(hd, temp);
+		free_append_word(head, temp);
 		return (NULL);
 	}
-	word_malloc(temp, srt, &n_srt, &end);
+	temp->lexeme = lexeme_malloc(temp, &i, start, &end);
 	trave = NULL;
-	trave = hd;
+	trave = head;
 	while (trave->next != NULL)
 		trave = trave->next;
 	trave->next = temp;
-	return (hd);
+	return (head);
 }
